:bug: update filterLevel type to number and refactor related components
