+++ src/entities/channel.d.ts
@@ -1,11 +1,10 @@
 type ChannelType = 'private' | 'group';
-type FilterType = 100 | 200 | 300;
 
 interface Channel {
   type: ChannelType;
   title: string;
   content: string;
-  filterLevel: FilterType;
+  filterLevel: number;
   link: string;
 }
