--- src/pages/channel/config/index.ts
@@ -1,5 +1,0 @@
-export const CHANNEL_ACTION_KEY = {
-  create: 'channel-create',
-  update: 'channel-update',
-  delete: 'channel-delete',
-};
