+++ src/pages/channel/store/form.ts
@@ -5,7 +5,7 @@ type State = ChannelCreateRequest;
 interface Actions {
   setTitle: (title: string) => void;
   setContent: (content: string) => void;
-  setFilterLevel: (filterLevel: FilterType) => void;
+  setFilterLevel: (filterLevel: number) => void;
   reset: () => void;
 }
 
@@ -15,10 +15,10 @@ export const useChannelFormStore = create<Store>((set) => ({
   type: 'private',
   title: '',
   content: '',
-  filterLevel: 100,
+  filterLevel: 0,
 
   setTitle: (title) => set({ title }),
   setContent: (content) => set({ content }),
   setFilterLevel: (filterLevel) => set({ filterLevel }),
-  reset: () => set({ title: '', content: '', filterLevel: 100 }),
+  reset: () => set({ title: '', content: '', filterLevel: 0 }),
 }));
